feat : 상품 목록 데이터 소스 추상화
